pyproject.toml
[tool.poetry]
name = "customer-care-paysoko"
version = "0.1.0"
description = ""
authors = ["Princekrampah <jsksprince@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
neo4j = "^5.27.0"
langchain-anthropic = "^0.3.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
langchain = "^0.3.11"
langchain-community = "^0.3.11"
fastapi = "^0.115.6"
uvicorn = "^0.33.0"
gradio = "^5.9.0"
streamlit = "^1.41.1"
numpy = ">=1.22.4,<2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
